+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.4.2] - 2019-10-09
+### Fixed
+- Fixed import for @required
+
+## [0.4.0] - 2019-10-09
+### Added
+- Named params to the models
+
 ## [0.3.0] - 2019-10-09
 ### Added
 - Nullable fields added